.gitignore and .gitattributes
We'll put art assets (binary, non-version-controlled assets) in
Assets/art/

.gitattributes is supposed to configure which file extensions are always
converted to LF on commit. So this means we have to change update this if
we commit a new text-format file extension. Stupid but I can't find a
better way to do it on a per-repo basis (otherwise we need to do it in our
per-user git config which is stupid)